PrimomoLanini-ADC return value

+++ ADC/ADC_prova_DATESTARE/src/include/adc.h
@@ -1,4 +1,4 @@
 #pragma once
 
 void adc_init(void);
-float adc_run(uint8_t, uint8_t);
+uint16_t adc_run(uint8_t, uint8_t);
